+++ handlers/user.py
@@ -17,6 +17,7 @@ class Handler(base.Handler):
 
     @gen.coroutine
     def post(self):
+        self.check_authorized()
         log.info('post user %s' % self.request.body)
         name = json.loads(self.request.body)['name']
         response = self.response
@@ -26,6 +27,7 @@ class Handler(base.Handler):
 
     @gen.coroutine
     def delete(self):
+        self.check_authorized()
         log.info('delete user %s' % self.request.body)
         id = json.loads(self.request.body)['id']
         self.twitter.delete_user(id)
@@ -34,6 +36,7 @@ class Handler(base.Handler):
 
     @gen.coroutine
     def get(self):
+        self.check_authorized()
         id = self.get_argument('id', None)
         log.info('get user %s' % id)
         response = self.response
